requirements.txt
openai
llama-index 
pydantic==1.10.14 
pillow
matplotlib
boto3
streamlit

